+++ web_app/node-server/src/graphql/TypeDefs/TypesTypeDefs.ts
@@ -90,9 +90,9 @@ const TypesTypeDefs = `
     gender: String
     profilePicture: String
     height: Int
-    weight: Int
+    weight: Float
     fitnessLevel: FitnessLevel
-    yearsOfExperience: Int
+    yearsOfExperience: Float
     primaryGoal: String
     secondaryGoals: [String]
     preferredWorkoutDays: Int
